# Separate buffer management and split point calculation