--- a/chunker.go
+++ b/chunker.go
@@ -35,14 +35,6 @@ func init() {
 	cache.entries = make(map[Pol]tables)
 }
 
-// Chunk is one content-dependent chunk of bytes whose end was cut when the
-// Rabin Fingerprint had the value stored in Cut.
-type Chunk struct {
-	Length uint
-	Cut    uint64
-	Data   []byte
-}
-
 type chunkerState struct {
 	window [windowSize]byte
 	wpos   uint
@@ -50,10 +42,6 @@ type chunkerState struct {
 
 	pre   uint // wait for this many bytes before start calculating an new chunk
 	count uint // used for max chunk size tracking
-
-	buf  []byte
-	bpos uint
-	bmax uint
 }
 
 type chunkerConfig struct {
@@ -64,13 +52,10 @@ type chunkerConfig struct {
 	tables            tables
 	tablesInitialized bool
 	splitmask         uint64
-
-	rd     io.Reader
-	closed bool
 }
 
 // Chunker splits content with Rabin Fingerprints.
-type Chunker struct {
+type BaseChunker struct {
 	chunkerConfig
 	chunkerState
 }
@@ -78,25 +63,19 @@ type Chunker struct {
 // SetAverageBits allows to control the frequency of chunk discovery:
 // the lower averageBits, the higher amount of chunks will be identified.
 // The default value is 20 bits, so chunks will be of 1MiB size on average.
-func (c *Chunker) SetAverageBits(averageBits int) {
+func (c *BaseChunker) SetAverageBits(averageBits int) {
 	c.splitmask = (1 << uint64(averageBits)) - 1
 }
 
-// New returns a new Chunker based on polynomial p that reads from rd.
-func New(rd io.Reader, pol Pol) *Chunker {
-	return NewWithBoundaries(rd, pol, MinSize, MaxSize)
+func NewBase(pol Pol) *BaseChunker {
+	return NewBaseWithBoundaries(pol, MinSize, MaxSize)
 }
 
-// NewWithBoundaries returns a new Chunker based on polynomial p that reads from
-// rd and custom min and max size boundaries.
-func NewWithBoundaries(rd io.Reader, pol Pol, min, max uint) *Chunker {
-	c := &Chunker{
-		chunkerState: chunkerState{
-			buf: make([]byte, chunkerBufSize),
-		},
+func NewBaseWithBoundaries(pol Pol, min, max uint) *BaseChunker {
+	c := &BaseChunker{
+		chunkerState: chunkerState{},
 		chunkerConfig: chunkerConfig{
 			pol:       pol,
-			rd:        rd,
 			MinSize:   min,
 			MaxSize:   max,
 			splitmask: (1 << 20) - 1, // aim to create chunks of 20 bits or about 1MiB on average.
@@ -104,25 +83,21 @@ func NewWithBoundaries(rd io.Reader, pol Pol, min, max uint) *Chunker {
 	}
 
 	c.reset()
-
 	return c
 }
 
 // Reset reinitializes the chunker with a new reader and polynomial.
-func (c *Chunker) Reset(rd io.Reader, pol Pol) {
-	c.ResetWithBoundaries(rd, pol, MinSize, MaxSize)
+func (c *BaseChunker) Reset(pol Pol) {
+	c.ResetWithBoundaries(pol, MinSize, MaxSize)
 }
 
 // ResetWithBoundaries reinitializes the chunker with a new reader, polynomial
 // and custom min and max size boundaries.
-func (c *Chunker) ResetWithBoundaries(rd io.Reader, pol Pol, min, max uint) {
-	*c = Chunker{
-		chunkerState: chunkerState{
-			buf: c.buf,
-		},
+func (c *BaseChunker) ResetWithBoundaries(pol Pol, min, max uint) {
+	*c = BaseChunker{
+		chunkerState: chunkerState{},
 		chunkerConfig: chunkerConfig{
 			pol:       pol,
-			rd:        rd,
 			MinSize:   min,
 			MaxSize:   max,
 			splitmask: (1 << 20) - 1,
@@ -132,7 +107,7 @@ func (c *Chunker) ResetWithBoundaries(rd io.Reader, pol Pol, min, max uint) {
 	c.reset()
 }
 
-func (c *Chunker) reset() {
+func (c *BaseChunker) reset() {
 	c.polShift = uint(c.pol.Deg() - 8)
 	c.fillTables()
 
@@ -140,7 +115,6 @@ func (c *Chunker) reset() {
 		c.window[i] = 0
 	}
 
-	c.closed = false
 	c.digest = 0
 	c.wpos = 0
 	c.count = 0
@@ -152,7 +126,7 @@ func (c *Chunker) reset() {
 
 // fillTables calculates out_table and mod_table for optimization. This
 // implementation uses a cache in the global variable cache.
-func (c *Chunker) fillTables() {
+func (c *BaseChunker) fillTables() {
 	// if polynomial hasn't been specified, do not compute anything for now
 	if c.pol == 0 {
 		return
@@ -205,67 +179,9 @@ func (c *Chunker) fillTables() {
 	cache.entries[c.pol] = c.tables
 }
 
-// Next returns the position and length of the next chunk of data. If an error
-// occurs while reading, the error is returned. Afterwards, the state of the
-// current chunk is undefined. When the last chunk has been returned, all
-// subsequent calls yield an io.EOF error.
-func (c *Chunker) Next(data []byte) (Chunk, error) {
-	data = data[:0]
-	for {
-		if c.bpos >= c.bmax {
-			n, err := io.ReadFull(c.rd, c.buf)
-
-			if err == io.ErrUnexpectedEOF {
-				err = nil
-			}
-
-			// io.ReadFull only returns io.EOF when no bytes could be read. If
-			// this is the case and we're in this branch, there are no more
-			// bytes to buffer, so this was the last chunk. If a different
-			// error has occurred, return that error and abandon the current
-			// chunk.
-			if err == io.EOF && !c.closed {
-				c.closed = true
-
-				// return current chunk, if any bytes have been processed
-				if len(data) > 0 {
-					return Chunk{
-						Length: uint(len(data)),
-						// somewhat meaningless as this is not a split point
-						Cut:  c.digest,
-						Data: data,
-					}, nil
-				}
-			}
-
-			if err != nil {
-				return Chunk{}, err
-			}
-
-			c.bpos = 0
-			c.bmax = uint(n)
-		}
-
-		split, cut := c.NextSplitPoint(c.buf[c.bpos:c.bmax])
-		if split == -1 {
-			data = append(data, c.buf[c.bpos:c.bmax]...)
-			c.bpos = c.bmax
-		} else {
-			data = append(data, c.buf[c.bpos:c.bpos+uint(split)]...)
-			c.bpos += uint(split)
-
-			return Chunk{
-				Length: uint(len(data)),
-				Cut:    cut,
-				Data:   data,
-			}, nil
-		}
-	}
-}
-
 // NextSplitPoint returns the index before which the buf should be split
 // Returns -1 if no split point was found yet.
-func (c *Chunker) NextSplitPoint(buf []byte) (int, uint64) {
+func (c *BaseChunker) NextSplitPoint(buf []byte) (int, uint64) {
 	if !c.tablesInitialized {
 		panic("tables for polynomial computation not initialized")
 	}
@@ -338,7 +254,7 @@ func updateDigest(digest uint64, polShift uint, tab *tables, b byte) (newDigest 
 	return digest
 }
 
-func (c *Chunker) slide(digest uint64, b byte) (newDigest uint64) {
+func (c *BaseChunker) slide(digest uint64, b byte) (newDigest uint64) {
 	out := c.window[c.wpos]
 	c.window[c.wpos] = b
 	digest ^= uint64(c.tables.out[out])
@@ -353,4 +269,126 @@ func appendByte(hash Pol, b byte, pol Pol) Pol {
 	hash |= Pol(b)
 
 	return hash.Mod(pol)
+}
+
+// Chunk is one content-dependent chunk of bytes whose end was cut when the
+// Rabin Fingerprint had the value stored in Cut.
+type Chunk struct {
+	Length uint
+	Cut    uint64
+	Data   []byte
+}
+
+type chunkerBuffer struct {
+	buf  []byte
+	bpos uint
+	bmax uint
+
+	rd     io.Reader
+	closed bool
+}
+
+// Chunker splits content with Rabin Fingerprints.
+type Chunker struct {
+	BaseChunker
+	chunkerBuffer
+}
+
+// New returns a new Chunker based on polynomial p that reads from rd.
+func New(rd io.Reader, pol Pol) *Chunker {
+	return NewWithBoundaries(rd, pol, MinSize, MaxSize)
+}
+
+// NewWithBoundaries returns a new Chunker based on polynomial p that reads from
+// rd and custom min and max size boundaries.
+func NewWithBoundaries(rd io.Reader, pol Pol, min, max uint) *Chunker {
+	c := &Chunker{
+		BaseChunker: *NewBaseWithBoundaries(pol, min, max),
+		chunkerBuffer: chunkerBuffer{
+			buf: make([]byte, chunkerBufSize),
+			rd:  rd,
+		},
+	}
+
+	c.reset()
+	return c
+}
+
+// Reset reinitializes the chunker with a new reader and polynomial.
+func (c *Chunker) Reset(rd io.Reader, pol Pol) {
+	c.ResetWithBoundaries(rd, pol, MinSize, MaxSize)
+}
+
+// ResetWithBoundaries reinitializes the chunker with a new reader, polynomial
+// and custom min and max size boundaries.
+func (c *Chunker) ResetWithBoundaries(rd io.Reader, pol Pol, min, max uint) {
+	c.BaseChunker.ResetWithBoundaries(pol, min, max)
+
+	*c = Chunker{
+		BaseChunker: c.BaseChunker,
+		chunkerBuffer: chunkerBuffer{
+			buf: c.buf,
+			rd:  rd,
+		},
+	}
+
+	c.closed = false
+}
+
+// Next returns the position and length of the next chunk of data. If an error
+// occurs while reading, the error is returned. Afterwards, the state of the
+// current chunk is undefined. When the last chunk has been returned, all
+// subsequent calls yield an io.EOF error.
+func (c *Chunker) Next(data []byte) (Chunk, error) {
+	data = data[:0]
+	for {
+		if c.bpos >= c.bmax {
+			n, err := io.ReadFull(c.rd, c.buf)
+
+			if err == io.ErrUnexpectedEOF {
+				err = nil
+			}
+
+			// io.ReadFull only returns io.EOF when no bytes could be read. If
+			// this is the case and we're in this branch, there are no more
+			// bytes to buffer, so this was the last chunk. If a different
+			// error has occurred, return that error and abandon the current
+			// chunk.
+			if err == io.EOF && !c.closed {
+				c.closed = true
+
+				// return current chunk, if any bytes have been processed
+				if len(data) > 0 {
+					return Chunk{
+						Length: uint(len(data)),
+						// somewhat meaningless as this is not a split point
+						Cut:  c.digest,
+						Data: data,
+					}, nil
+				}
+			}
+
+			if err != nil {
+				return Chunk{}, err
+			}
+
+			c.bpos = 0
+			c.bmax = uint(n)
+		}
+
+		split, cut := c.NextSplitPoint(c.buf[c.bpos:c.bmax])
+		if split == -1 {
+			data = append(data, c.buf[c.bpos:c.bmax]...)
+			c.bpos = c.bmax
+		} else {
+			data = append(data, c.buf[c.bpos:c.bpos+uint(split)]...)
+			c.bpos += uint(split)
+
+			return Chunk{
+				Length: uint(len(data)),
+				Cut:    cut,
+				Data:   data,
+			}, nil
+		}
+	}
 }