`Chunker.Next` currently handles buffer management and the split point calculation at the same time. While experimenting with a way to also chunk Tree blobs, I noticed that this combination is a problem: trees are built incrementally such that the Chunker cannot simply read the next part of the tree into its buffer. Moving the buffer out of the Chunker helps in this case, as it allows to manage a buffer in a way that better matches what is necessary to split tree blobs.

For now the PR splits the chunker into `BaseChunker` and `Chunker`. However, if the tree blob chunking works out as planned, then the `Chunker` will no longer be used by restic and could be removed unless there are other users.

Performance seems to be mostly unaffected, maybe even improve a tiny bit:
```
name               old time/op   new time/op   delta
ChunkerWithSHA256   93.3ms ± 0%   93.2ms ± 0%    ~     (p=1.000 n=9+9)
Chunker             41.7ms ± 0%   41.5ms ± 1%  -0.41%  (p=0.002 n=9+9)

name               old speed     new speed     delta
ChunkerWithSHA256  360MB/s ± 0%  360MB/s ± 0%    ~     (p=0.950 n=9+9)
Chunker            805MB/s ± 0%  808MB/s ± 1%  +0.41%  (p=0.002 n=9+9)
```